New version of clang (19) has introduced a stricter checks when mixing
integer (and float types) with enums.  In this case, we used enum {}
as C17 doesn't have constexpr yet.  Change the time conversion constants
to be #defined constants because of RHEL 8 compiler doesn't consider
static const unsigned int to be constant.

Closes #4845

Backport of MR !9313

Merge branch 'backport-4845-change-NS_PER_SEC-type-from-enum-to-integer-9.18' into 'bind-9.18'

See merge request isc-projects/bind9!9340

+++ lib/isc/stdtime.c
@@ -42,7 +42,8 @@ isc_stdtime_get(isc_stdtime_t *t) {
 		FATAL_SYSERROR(errno, "clock_gettime()");
 	}
 
-	REQUIRE(ts.tv_sec > 0 && ts.tv_nsec >= 0 && ts.tv_nsec < NS_PER_SEC);
+	REQUIRE(ts.tv_sec > 0 && ts.tv_nsec >= 0 &&
+		ts.tv_nsec < (long)NS_PER_SEC);
 
 	*t = (isc_stdtime_t)ts.tv_sec;
 }
